refactor(pr06): rename applyIdentifiableSchema.mts -> IdentifiableSchemaMixin.mts (G5c)
File-only rename to match exported symbol IdentifiableSchemaMixin.
