+++ src/entities/components/Identifiable/data/index.mts
@@ -1,8 +1,8 @@
-import { IdentifiableSchemaMixin } from './applyIdentifiableSchema.mjs';
 import type {
   IdentifiableDocumentSystemData,
   IdentifiableDocumentSystemSource,
 } from './IdentifiableDocumentSystemData.mjs';
+import { IdentifiableSchemaMixin } from './IdentifiableSchemaMixin.mjs';
 
 export {
   IdentifiableSchemaMixin,
